requirements.txt
alembic==1.7.6
boto3==1.20.49
Flask==2.0.2
Flask_Bootstrap==3.3.7.1
Flask_HTTPAuth==4.5.0
Flask_Login==0.5.0
Flask_Mail==0.9.1
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.0
PyJWT==2.3.0
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1
